auto installer: make fetch  print answer to stdout and explicitly run installer
Instead of executing the actual installer from inside a tool named
`fetch-answer` make it print the answer file content it queried to
stdout and use the outer "unconfigured" init process to redirect this
to a file inside run, which is then piped to the actual installer in a
separate command.

This makes the execution flow a bit easier to grasp when reading
unconfigured and allows to inspect the answer file in a debug session.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>