auto installer: rework sys-info struct layout and add more details
move the decoded DMI info into a "dmi" object to avoid crowding the
outer layer, add more info from the ISO and Product currently booted
and use the complete mocked variants for the assistant environments so
that users evaluating this have a full picture of possible keys.

Also include the NIC link name with the mach address so that one can
also configure the management interface correctly from a MAC.

Do this all in a slightly more ergonomic rust way, using separate
structs and impls for getting the info.

There could be even more details added, but that's for a future patch,
this one mostly transforms and enhances existing keys.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>