assistant: rework prepare-iso command line interface
Use input and output for the input and output files, drop various
short options, we can always re-add them later and forcing the use of
the more telling long options is a UX gift to sites with multiple
admins working on this stuff.

Further rename the "install-mode" to "fetch-from" as the option does
not changes the installation mode but rather the mode where the answer
file is fetched from.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>