assistant: default to output directory for tmp iso file on prepare
The directory where the input ISO is on might be read-only and so it
makes more sense to use the output directory to derive the default tmp
directory from.

Note that by default the output directory is the same as the one from
the input file, so this commit won't have a change for those that do
not override the output file explicitly.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>