fetch answer: allow to override fetch-from mode through CLI arguments
Allow the user to pass a fetch-from mode also through CLI arguments.

This can be useful for debugging or if the GRUB boot entry from the
advanced menu is used with a ISO that did not (yet) got prepared for
automatic installation.

This is done in a quite crudely way, probably still beats most C
programs though. Only real bigger ugliness the user will notice is
that it exits with a failure code when the user asking for help
through passing `-h` or `--help`.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>