tui: switch cursive to crossterm backend
Development of crossterm seems to be more active, has more frequent
releases and is also the default backend for cursive - so overall also
probably better supported/tested.

Additionally, it feels a bit more snappier/less flickery, probably due
to buffering writes to the TTY layer unlike termion.

As this effects only the way cursive talks to the TTY, this has no
visible change on the TUI itself.

Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>