fetch-answer: add `$fetchinfo` meta-field to HTTP POST payload
This adds a metadata-field `$fetchinfo` containing a single key
`version` (for now) to the POST payload json, indicating which schema
version (and thus structure) this document uses.

The version field follows the format "<major>.<minor>" and applies
semantic versioning meaning for both the major and minor number. A patch
version is left out here, as it doesn't make much sense in this context.

Works in the same manner as the post-hook does it [0]. Useful to have it
for this too, as we might change/expand this structure too in the
future.

In the resulting JSON, this will look like this:
{
    "$fetchinfo": {
	"version": "1.0"
    },
    "product": ..,
    ..
}

[0] https://lore.proxmox.com/pve-devel/20241112145405.929194-1-c.heiss@proxmox.com/

Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>