auto-installer: fix: add default value for partition-label
This ensure backwards-compatibility of older version of
proxmox-auto-install-assistant and newer ISOs/auto-installer version.

Otherwise, it fails with a "missing key `partition_label`" error on
start.

While proxmox-auto-install-assistant writes out that field by default
now, it will be missing for ISOs prepared with older versions of that
tool.

Fix it by providing a default label, such that it always has a sensible
value.

Fixes: 2e89c07 ("auto-install-assistant: add new parameter to specify partition label")
Reported-by: Timothy Nicholson <t.nicholson@proxmox.com>
Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>