early init: set kernel.printk to level warn
we got a few reports of kernel-output drawing over the TUI-installer.
(e.g. when ZFS loads for the pool-cleanup, also mkfs.xfs sometimes
triggers a message being printed).

As the information gets logged in the kernel ringbuffer (`dmesg`)
anyways, there's little gain on having it displayed on the console in
addition.

Following the kernel docs [0] and a guide from Red Hat [1] simply set
the sysctl to log only messages of level warn and above to the
console.

This got tested in a VM, where I manually ran the command before
starting the TUI installer (and selecting XFS as filesystem there).
without it the messages got displayed, with the command they did not.

[0]: https://www.kernel.org/doc/Documentation/sysctl/kernel.txt
[1]: https://docs.redhat.com/en/documentation/red_hat_enterprise_linux/8/html/managing_monitoring_and_updating_the_kernel/getting-started-with-kernel-logging_managing-monitoring-and-updating-the-kernel

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>