fix #2502: gtk ui: network settings: set CIDR text manually
The installer commits user-specified changes made in a wizard step
when it advances to the next step, otherwise it should keep the
detected defaults, or if none could be detected, or if that's not
applicable, some hardcoded default.

But there was an edge case with the previous/back handling if you
never progressed beyond the Network Settings wizard step. Because
there we only applied the detected IP address from DHCP the first
time, but not if you went back one or more steps from the Network step
and then went back to the Network step.

To fix this, drop the special handling for displaying the network
config step the first time by triggering the device_change_handler
callback (which saves the config) and instead manually set the
selected NIC and the address and mask text to the correct value.

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
 [TL: expand commit message with more background]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>