auto-installer: allow snake- and kebabcased property names in answer files
The default case for property names in TOML configuration files is de
facto snake_case. The answer file for the auto-installer adapted this,
but this causes inconsistencies with the property names in other Proxmox
config file formats, which are preferably in kebab-case.

The properties for the "first-boot" section were intentionally
introduced in kebab-case to steer away from snake_cased to kebab-cased
property names. Since this caused some initial confusion for users [0],
allow both cases for all properties for the time being and deprecate
the snake_cased variant of all properties in a future major version.

[0] https://bugzilla.proxmox.com/show_bug.cgi?id=5973

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>