run env: always leave 1 GiB for system when clamping ZFS ARC size
To not allow the max ARC size to be the same as the total memory size.
Note that 1 GiB is often still not enough for a fully working system,
but for one, this can only happen if the user manually overrides ARC
max, and for another, it's only a general crude safety net to allow
the basic system to boot so that one can correct this if it's a
problem as for some products, like PBS or PDM, the base system's RSS
usage might be below 1 GiB, and we are not doing product-aware stuff
here.

Adapt the tests accordingly.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>