proxmox-chroot: replace nix::(u)mount calls with external (u)mount
Comes with a reduction of 52 -> 40 in terms of crate dependencies for
proxmox-chroot, 198 -> 192 for a full workspace build.

Currently, this is done inconsistently anyway, i.e. there are calls to
the external mount(8) as well as mount(2) and umount(2) via `nix`.
Just switch over to calling the external programs completely, which in
turn allows to drop the `nix` crate dependency from the tree.

No functional changes.

Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>