proxinstall: replace usage of hacky `mngmt_nic_id` config option
The option was more or less a hack from the beginning anyway, introduced
during the split of the GUI and the installation configuration.

The required information is already there from `mngmt_nic` - which
carries the name of the network interface - and is easily reused to
replace the functionality.

In short, the `mngmt_nic_id` setting was only used to check whether
another NIC was selected, avoiding needless updates to the IP address
and mask textboxes.

Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>