gui: fall back to first found NIC if no default could be determined
If no DHCP server is configured on the network and/or no DHCP lease and
thus network configuration is retrieved, the installer will also be
unable to determine the default interface, aka. the one with the default
gateway set.

Fix it by pre-selecting the first found interface. If we *do* have a
default interface, the index will be later overwritten, but ensures that
always *some* interface is pre-selected in the dropdown.

As it is already checked that the machine has any network interfaces
before starting the installation wizard, it is safe to rely on the fact
that at least one NIC must be present.

For completeness, all the other network inputs already have (hardcoded)
fallbacks, namely 192.168.100.2/24 as address/netmask, 192.168.100.1 as
gateway and the gateway address as DNS server address.

Reported-by: Christian Ebner <c.ebner@proxmox.com>
Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>
Link: https://lore.proxmox.com/20250407135510.1138958-1-c.heiss@proxmox.com