btrfs: pass UUID of root file system to sd-boot via kernel cmdline
The recent change of handing the ESPs to proxmox-boot-tool broke
installation on btrfs booted with systemd-boot (currently efi-systems
without secure boot). our systemd-boot config-generation takes the
kernel command line from /etc/kernel/cmdline - without this there is
no root= parameter, and we end up in a initramfs shell.

Write the file unconditionally of boot-mode (legacy, efi,
efi-secureboot), to be consistent with the way we do it on ZFS.

The UUID is stored in a variable instead of running blkid multiple
time primarily to not end up with different versions of a similar
command in the future, at the expense of 3 change-sites which are
quite far apart.

Fixes: dc5b8d7 ("btrfs: use proxmox-boot-tool to sync ESPs for RAID installs")
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20250408191806.2410107-1-s.ivanov@proxmox.com
 [TL: rework commit message subject]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>