assistant: replace clap with pico-args for command argument parsing
Drops clap for pico-args, as the former is pretty heavy-weight and
completely overhauls its API with ~every major release.

While it results in a bit more boilerplate code, it makes (especially
long-term) maintenance simpler.

As for the resulting binary, there is a ~13.5% decrease in binary size,
from 4M to 3.4MiB (after stripping).

   text    data     bss     dec     hex filename
3793413  299000     616 4093029  3e7465 proxmox-auto-install-assistant-clap
3245990  286136     472 3532598  35e736 proxmox-auto-install-assistant-pico

Further, the dependency tree goes from 114 to 103 dependencies, as well
as the package compile time on my machine (on hot caches), roughly
measured with

`cargo clean && time cargo build --package proxmox-chroot --release`

from ~31 to ~28s.

A big chunk of the added lines are simply the help menus for the
different subcommands of the tool, the actual code changes are rather
small.

No functional changes intended.

Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>
Link: https://lore.proxmox.com/20250509121007.1430080-4-c.heiss@proxmox.com