common: initial upgrade to ureq 3 and rustls 0.23
incomplete as TLS cert validation needs to be fixed, we now need to
implement our own transport to be able to pass a rustls ClientConfig,
see an example here [0]. FWIW, ureq is evaluating if the verification
for cert fingerprints is important enough for ureq's main focus and
use cases to have better native support [1], so this has some chance
to be come simpler again in the future.

but update the rest to be prepared for that already.

[0]: https://github.com/algesten/ureq/pull/1085/files
[1]: https://github.com/algesten/ureq/issues/1087#issuecomment-2984113154

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>