common: http: add custom ureq connector for cert-fingerprint validation
Adds back custom TLS fingerprint validation for ureq 3. The API was
changed majorly, no longer providing the possibility to directly provide
a rustls `ClientConfig` to the agent.

Instead, a complete custom transport + connector must be provided to
ureq to achieve this.

Mostly based on ureq#1085 [0], which provides a example how to implement
something like this.

[0] https://github.com/algesten/ureq/pull/1085

Fixes: 2557dbf ("common: initial upgrade to ureq 3 and rustls 0.23")
Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>
Link: https://lore.proxmox.com/20250625122229.1115619-1-c.heiss@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>