common: cli: always print trailing newline on command error
most our sub-commands use just bail! on error, which would result in
the error message having no trailing newline and thus being "glued" to
the start of the shell prompt. Always output a newline, in the worst
case a duplicate newline results in the error message sticking out
even more.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>