run-env: filter out interfaces without valid mac address
The installer assumes to have a valid mac address for all interfaces
as provided by the runtime env json file. Deserialization will fail
if this is not the case.

In some cases, the interface might however not provide a valid MAC
address, for example the WWAN LTE module without any SIM installed on
some laptops.

Filter out these interfaces and log this, as they cannot be correctly
used anyways.

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250711090035.241160-1-c.ebner@proxmox.com