run env: anchor domain suffix strip regex to end of string
While it is very much an edge case, properly anchoring the regex ensures
that the local domain only ever gets stripped if it is indeed a suffix
of the fully-qualified hostname.

Prevents situations where the DHCP server sends the following client
configuration:

  host-name: host.foo.bar
  domain-name: foo

Which would result in `bar.foo` being set as the search domain in the
final system without this patch.

Fixes: a9c0a44 ("run env: fix dhcp-set hostname containing local domain")
Signed-off-by: Christoph Heiss <c.heiss@proxmox.com>
Link: https://lore.proxmox.com/20250716101320.571502-1-c.heiss@proxmox.com