gui: don't assume 'single' btrfs mode if user selects exactely 1 disk
assuming 'single' mode if the user selects exactely one disk can lead
to unexpected results. for example, if a user sets up a raid10 with
one disk, the installer will happily install it as a 'single' install.
doing the same with two disks fails with an error that for raid10 at
least 4 disks are required.

this is also inconsistent with the tui installer that will always
complain about the amount of disks not being correct.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250716132557.237727-1-s.sterz@proxmox.com