install: set timezone, before configuring postfix
quite a few changes have been made to the postfix package in trixie.
one of those is that `configure-instance.sh`, which handles
chroot-creation is now called on service startup and e.g. with
`postfix check`, which we run here[0].

this results in /etc/localtime being copied into the changeroot
before we set it to the timezone configured by the installer,
resulting in a warning upon rebooting (which would remain
until you either `dpkg-reconfigure tzdata` or `touch /etc/localtime`.

[0] https://salsa.debian.org/postfix-team/postfix-dev/-/commit/6a3cfe2931138898bea99b171c3731b17ee2fbae

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Tested-by: Hannes Duerr <h.duerr@proxmox.com>
Link: https://lore.proxmox.com/20250805213629.1108359-1-s.ivanov@proxmox.com