install: do not create deprecated /etc/timezone
starting with  tzdata version 2024b-5 [0] the package does not create
/etc/timezone anymore (background is quite well explained in
/usr/share/doc/tzdata/NEWS.Debian.gz:
```
Previously, the tzdata package in Debian used the /etc/timezone file to
configure the system's timezone. This method is not supported by systemd
and certain desktop environments, which instead only change the
/etc/localtime symlink to point to a file in /usr/share/zoneinfo.
....
In a future release, the /etc/timezone file will be automatically
removed ... The debian-installer from Trixie also no longer
creates this file.
```
align our installer with debian's installer.

[0] https://metadata.ftp-master.debian.org/changelogs//main/t/tzdata/tzdata_2025b-4_changelog

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20250805214006.1111543-1-s.ivanov@proxmox.com