d/rules: allow for local testing suffixes to debian versions
When testing the installer in a ISO environment it can be quite
helpful to change the version of the debian-package to a temporary
higher version (which still orders below the next released version).

e.g. I now wanted to build a installer-package for some tests with
debian version 9.0.10~localtest1, but the build failed at
dh_auto_configure due to the mismatch.

While disabling/removing the check in those cases is not too much
effort either it might help others as well in not running into this.

I've only allowed for `~` and `+` as strippable suffix, as `-` would
at least cause lintian to complain about a strange version for a
native package.

Background on debian package versions:
https://www.debian.org/doc/debian-policy/ch-controlfields.html#version

This a change to
9528321 ("d/rules: add safe-guard to keep cargo and debian version in sync")

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251106125658.3362672-1-s.ivanov@proxmox.com